Outputs as an include to prevent duplicate code
