+++ src/pages/includes/multiple-outputs/multiple-outputs.html
@@ -1,0 +1,57 @@
+<ion-item class="container-to" (click)="showMultiplesOutputs = !showMultiplesOutputs">
+  <span translate>To</span>
+  <div class="wallet">
+    <div class="address-coppied" [ngClass]="{'address-width':tx.hasMultiplesOutputs}">
+      <div class="text-address ellipsis">
+        <img *ngIf="tx.hasMultiplesOutputs || (!tx.hasMultiplesOutputs && tx.customData?.service != 'debitcard')" src="assets/img/icon-bitcoin-small.svg" width="12">
+        <img *ngIf="!tx.hasMultiplesOutputs && tx.customData?.service == 'debitcard'" src="assets/img/icon-card.svg" class="debit-card">
+        <span class="address-divider"></span>
+      </div>
+
+      <div class="ellipsis" *ngIf="!tx.hasMultiplesOutputs" copy-to-clipboard="{{ tx.toAddress ? tx.toAddress : tx.outputs[0].address }}">
+        <span *ngIf="!contactName && !tx.toAddress">{{ tx.outputs[0].address }}</span>
+        <span *ngIf="!contactName && tx.toAddress">{{ tx.toAddress }}</span>
+        <span *ngIf="contactName">{{ contactName }}</span>
+      </div>
+
+      <div class="ellipsis" *ngIf="tx.hasMultiplesOutputs">
+        <span>
+          <span translate>Multiple recipients: </span>
+          <span>{{tx.recipientCount}}</span>
+        </span>
+      </div>
+    </div>
+
+    <div class="last-item" *ngIf="tx.hasMultiplesOutputs" item-end>
+      <button ion-button clear color="grey" icon-only>
+        <ion-icon *ngIf="showMultiplesOutputs" name="ios-arrow-up-outline"></ion-icon>
+        <ion-icon *ngIf="!showMultiplesOutputs" name="ios-arrow-down-outline"></ion-icon>
+      </button>
+    </div>
+  </div>
+</ion-item>
+
+<ion-list detail-none *ngIf="tx.hasMultiplesOutputs && showMultiplesOutputs">
+  <ion-item class="proposal-container" *ngFor="let output of tx.outputs" copy-to-clipboard="{{output.toAddress ? output.toAddress : output.address}}">
+    <div item-left>
+      <h2>
+        <span translate>To</span>:
+      </h2>
+      <h3>
+        <span translate>Amount</span>:
+      </h3>
+      <p class="output-note" *ngIf="output.message">
+        <span translate>Note</span>:
+      </p>
+    </div>
+    <ion-note item-right text-end col-11>
+      <h2 class="output-info ellipsis">{{output.toAddress || output.address}}</h2>
+      <h3 class="output-info ellipsis">{{output.amountStr}}
+        <span *ngIf="output.alternativeAmountStr">({{output.alternativeAmountStr}})</span>
+      </h3>
+      <p class="output-note" *ngIf="output.message">
+        <span class="ellipsis">{{output.message}}</span>
+      </p>
+    </ion-note>
+  </ion-item>
+</ion-list>
