+++ src/pages/includes/multiple-outputs/multiple-outputs.scss
@@ -1,0 +1,30 @@
+page-multiple-outputs {
+  .address-coppied {
+    display: flex;
+    align-items: center;
+    width: 100%;
+    border-radius: 3px;
+    background-color: rgb(248, 248, 248);
+    .text-address {
+      flex: 0 0 47px;
+      img {
+        margin: 12px 0 12px 12px;
+        vertical-align: middle;
+      }
+      .debit-card {
+        width: 30px;
+        margin: 4px -6px 4px 6px;
+      }
+    }
+  }
+  .last-item {
+    margin-left: auto !important;
+    margin-right: 0px !important;
+  }
+  .output-info {
+    color: color($colors, dark);
+  }
+  .output-note {
+    color: color($colors, light-grey);
+  }
+}
