+++ src/pages/txp-details/txp-details.scss
@@ -71,20 +71,6 @@ page-txp-details {
     .wallet,
     .payment-proposal-to {
       justify-content: space-between;
-      .text-address.ellipsis {
-        flex: 0 0 47px;
-        img {
-          margin: 12px 0 12px 12px;
-          vertical-align: middle;
-        }
-      }
-      .address-coppied {
-        display: flex;
-        align-items: center;
-        width: 100%;
-        border-radius: 3px;
-        background-color: rgb(248, 248, 248);
-      }
       .address-width {
         width: 92%;
       }
@@ -207,19 +193,6 @@ page-txp-details {
         color: #aaa;
         font-size: 14px;
       }
-    }
-  }
-  .last-item {
-    margin-left: auto !important;
-    margin-right: 0px !important;
-  }
-  .output-info {
-    color: color($colors, dark);
-    text-overflow: ellipsis;
-    overflow: hidden;
-    white-space: nowrap;
-    &.output-note {
-      color: color($colors, grey);
     }
   }
   .proposal-container {
